add single pane view logic to FluentTesterApp
