+++ apps/fluent-tester/src/FluentTester/FluentTester.tsx
@@ -3,10 +3,10 @@ import { StealthButton, Separator } from '@fluentui/react-native';
 import { Text } from '@fluentui-react-native/experimental-text';
 import { themedStyleSheet } from '@fluentui-react-native/themed-stylesheet';
 import * as React from 'react';
-import { ScrollView, View, Text as RNText } from 'react-native';
+import { ScrollView, View, Text as RNText, BackHandler } from 'react-native';
 import { TestDescription } from './TestComponents';
 import { BASE_TESTPAGE } from './TestComponents/Common/consts';
-import { fluentTesterStyles } from './TestComponents/Common/styles';
+import { fluentTesterStyles, mobileStyles } from './TestComponents/Common/styles';
 import { useTheme } from '@fluentui-react-native/theme-types';
 import { ThemePickers } from './theme/ThemePickers';
 
@@ -20,8 +20,11 @@ const EmptyComponent: React.FunctionComponent = () => {
   return <RNText style={fluentTesterStyles.noTest}>Select a component from the left.</RNText>;
 };
 
+const DisplayIfVisible = ({ isVisible, children }) => (isVisible ? <View>{children}</View> : null);
+
 export interface FluentTesterProps {
   initialTest?: string;
+  enableSinglePaneView?: boolean;
   enabledTests: TestDescription[];
 }
 
@@ -64,12 +67,19 @@ export const FluentTester: React.FunctionComponent<FluentTesterProps> = (props: 
   // sort tests alphabetically by name
   const sortedTestComponents = props.enabledTests.sort((a, b) => a.name.localeCompare(b.name));
 
-  const { initialTest } = props;
+  const { initialTest, enableSinglePaneView } = props;
   const initialSelectedTestIndex = sortedTestComponents.findIndex((description) => {
     return description.name === initialTest;
   });
 
   const [selectedTestIndex, setSelectedTestIndex] = React.useState(initialSelectedTestIndex);
+  const [onTestListView, setOnTestListView] = React.useState(true);
+
+  const onBackPress = () => {
+    setOnTestListView(true);
+    BackHandler.removeEventListener('hardwareBackPress', onBackPress);
+    return true;
+  };
 
   const TestComponent = selectedTestIndex == -1 ? EmptyComponent : sortedTestComponents[selectedTestIndex].component;
 
@@ -86,30 +96,69 @@ export const FluentTester: React.FunctionComponent<FluentTesterProps> = (props: 
 
       <Separator />
 
-      <View style={fluentTesterStyles.testRoot}>
-        <ScrollView style={fluentTesterStyles.testList} contentContainerStyle={fluentTesterStyles.testListContainerStyle}>
-          {sortedTestComponents.map((description, index) => {
-            return (
-              <StealthButton
-                key={index}
-                disabled={index == selectedTestIndex}
-                content={description.name}
-                onClick={() => setSelectedTestIndex(index)}
-                style={fluentTesterStyles.testListItem}
-                testID={description.testPage}
-              />
-            );
-          })}
-        </ScrollView>
-
-        <TestListSeparator vertical style={{ marginHorizontal: 8, width: 2 }} />
-
-        <View style={fluentTesterStyles.testSection}>
-          <ScrollView>
-            <TestComponent />
-          </ScrollView>
+      {enableSinglePaneView ? (
+        <View>
+          <DisplayIfVisible isVisible={onTestListView}>
+            <View style={themedStyles.root}>
+              <ScrollView style={mobileStyles.testList} contentContainerStyle={fluentTesterStyles.testListContainerStyle}>
+                {sortedTestComponents.map((description, index) => {
+                  return (
+                    <View key={index}>
+                      <Text
+                        key={index}
+                        onPress={() => {
+                          setOnTestListView(false);
+                          setSelectedTestIndex(index);
+                          BackHandler.addEventListener('hardwareBackPress', onBackPress);
+                        }}
+                        style={mobileStyles.testListItems}
+                        testID={description.testPage}
+                      >
+                        {description.name}
+                      </Text>
+                      <Separator style={mobileStyles.testListSeparator} />
+                    </View>
+                  );
+                })}
+              </ScrollView>
+            </View>
+          </DisplayIfVisible>
+          <DisplayIfVisible isVisible={!onTestListView}>
+            <View style={themedStyles.root}>
+              <View style={mobileStyles.testSection}>
+                <ScrollView>
+                  <TestComponent />
+                </ScrollView>
+              </View>
+            </View>
+          </DisplayIfVisible>
         </View>
-      </View>
+      ) : (
+        <View style={fluentTesterStyles.testRoot}>
+          <ScrollView style={fluentTesterStyles.testList} contentContainerStyle={fluentTesterStyles.testListContainerStyle}>
+            {sortedTestComponents.map((description, index) => {
+              return (
+                <StealthButton
+                  key={index}
+                  disabled={index == selectedTestIndex}
+                  content={description.name}
+                  onClick={() => setSelectedTestIndex(index)}
+                  style={fluentTesterStyles.testListItem}
+                  testID={description.testPage}
+                />
+              );
+            })}
+          </ScrollView>
+
+          <TestListSeparator vertical style={{ marginHorizontal: 8, width: 2 }} />
+
+          <View style={fluentTesterStyles.testSection}>
+            <ScrollView>
+              <TestComponent />
+            </ScrollView>
+          </View>
+        </View>
+      )}
     </View>
   );
 };
